renaming file and changing respective file links
